Login works, settings, and loading screen
Added login functionality. Also added loading screens while auth operations are in effect. Created a cog in the top right corner of the generate screen with "account settings" and "logout". Account settings not yet implemented and logging out causes the app to crash with no error messages... Need to fix.